image_alteration/requirements.txt
# Image Processing Scripts Requirements
# Install with: pip install -r requirements.txt

# Core image processing
Pillow>=9.0.0

# RAW image processing
rawpy>=0.17.0

# Additional image I/O support
imageio>=2.19.0

# Numpy for advanced image operations
numpy>=1.21.0

# GUI framework (usually included with Python)
# tkinter - included with most Python installations

# Optional: Additional image format support
# opencv-python>=4.5.0  # Uncomment for OpenCV support
# scikit-image>=0.19.0  # Uncomment for advanced image processing